requirements.txt
altgraph==0.17.2
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==4.4.2
future==0.18.2
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.0
Pillow==9.1.1
pkg_resources==0.0.0
proglog==0.1.10
pyinstaller-hooks-contrib==2022.8
python-slugify==6.1.2
pytube==12.1.0
requests==2.28.1
six==1.16.0
text-unidecode==1.3
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.10
wxPython==4.1.1
